PXC-2560 : pxc_strict_mode failing intermittently on PXC-5.7.26 branch
Issue
The test was failing due to our setting a variable, but the statement
propagation was taking longer than expected (so the local set statement
would execute before some of the commands on the other node had
propagated).  This happens because the SET command does not replicate
and there is no sync wait for the SET statements.

Solution
Have the SET command (for global variables) also do a sync wait
if so enabled (we are extending the SHOW sync wait to SET).