PXC-3625: PXC: remove workaround code in SST script
Issue
There is code in the wsrep_sst_xtrabackup-v2 scripts that manually
move the binlog/binlog-index files.  This confuses xtrabackup,
since it is also trying to move the binlog files, and causes
an error during the move-back operation for the SST.

Solution
Remove the old code that moves the binlog-related files.